getting info for produt and user registeration apis working fine
