--- termic_backend/base/urls.py
@@ -1,40 +1,0 @@
-"""
-URL configuration for termic_backend project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
-from django.urls import path
-from . import views
-
-
-urlpatterns = [
-    path('users/login', views.MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
-    path('users/register/', views.registerUser, name='register'),
-    path('products/', views.getProducts, name='products'),
-    path('products/<str:pk>', views.getProduct, name='product'),
-    path('users/profile/', views.getUserProfile, name='users-profile'),
-    path('users/', views.getUsers, name='users'),
-
-
-
-
-
-
-
-
-
-
-]
-
