+++ termic_backend/base/urls/orders_urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from base.views import order_views as views
+
+
+urlpatterns = [
+    
+
+]
