+++ termic_backend/base/views/order_views.py
@@ -1,0 +1,1 @@
+pass
