+++ termic_backend/base/views/user_views.py
@@ -1,10 +1,7 @@
-from django.shortcuts import render
-from django.http import HttpResponse,JsonResponse
-from .products import products
 from rest_framework.decorators import api_view,permission_classes
 from rest_framework.response import Response
-from .models import Product,User
-from .serializers import ProductSerializer,UserSerializer,UserSerializerWithToken
+from base.models import User
+from base.serializers import UserSerializer,UserSerializerWithToken
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
@@ -12,10 +9,6 @@ from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from django.contrib.auth.hashers import make_password
 from rest_framework import status
 # Create your views here.
-
-
-
-
 
 
 class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
@@ -73,68 +66,6 @@ def getUsers(request):
     users = User.objects.all()
     serializer = UserSerializer(users,many =  True)
     return Response(serializer.data)
-
-
-
-
-
-
-
-@api_view(['GET'])
-def getProducts(request):
-    products = Product.objects.all()
-    serializer = ProductSerializer(products,many =  True)
-    return Response(serializer.data)
-
-
-
-
-
-@api_view(['GET'])
-def getProduct(request,pk):
-    # print('the pk is '+pk)
-    product = Product.objects.get(_id=pk)
-    serializer = ProductSerializer(product,many =  False)
-    return Response(serializer.data)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
