+++ termic_backend/termic_backend/urls.py
@@ -22,7 +22,12 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/',include('base.urls')),
+    # path('api/',include('base.urls')),
+    path('api/products/',include('base.urls.product_urls')),
+    path('api/users/',include('base.urls.user_urls')),
+    path('api/orders/',include('base.urls.orders_urls')),
+
+
 ]
 
 urlpatterns += static(settings.MEDIA_URL,document_root=settings.MEDIA_ROOT)
